+++ api/internal/konfig/builtinpluginconsts/images.go
