+++ api/internal/konfig/builtinpluginconsts/namesuffix.go
